requirements.txt
gunicorn==20.1.0
django-heroku==0.3.1
django-ckeditor==6.0.0
pillow==8.1.2
